Refactoring savvy app
