--- superset/savvy/savvymodels.py
@@ -1,44 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-# pylint: disable=C,R,W
-import datetime
-
-from flask_appbuilder import Model
-from flask_appbuilder.security.views import UserDBModelView
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, \
-    Sequence, String
-
-
-class ResetRequest(Model):
-    ___tablename__ = 'reset_request'
-    id = Column(Integer, Sequence('reset_request_id_seq'), primary_key=True)
-    user_id = Column(Integer, ForeignKey('ab_user.id'))
-    email = Column(String(64), nullable=False)
-    reset_date = Column(DateTime, default=datetime.datetime.now, nullable=True)
-    reset_hash = Column(String(256))
-    used = Column(Boolean)
-
-
-class SavvyUserDBModelView(UserDBModelView):
-
-    def pre_delete(self, user):
-        print(user)
-        organization = self.appbuilder.sm.find_org(user_id=user.id)
-        if len(organization.users) == 1:
-            self.appbuilder.sm.delete_org(organization)
-
-
