+++ superset/savvy/utils.py
@@ -1,0 +1,5 @@
+import json
+
+def post_request(url, params):
+    import requests
+    return requests.post(url, data=json.dumps(params))
