+++ superset/security.py
@@ -101,15 +101,6 @@ OBJECT_SPEC_PERMISSIONS = set([
 
 
 class SupersetSecurityManager(SecurityManager):
-    registeruserdbview = SavvyRegisterUserDBView
-
-
-    from superset.savvy.organization import OrgRegisterUser
-    registeruser_model = OrgRegisterUser
-
-    def register_views(self):
-        super(SecurityManager, self).register_views()
-        self.appbuilder.add_view_no_menu(SavvyRegisterInvitationUserDBView())
 
     def get_schema_perm(self, database, schema):
         if schema:
@@ -467,38 +458,3 @@ class SupersetSecurityManager(SecurityManager):
                 self.get_datasource_access_error_msg(datasource),
                 self.get_datasource_access_link(datasource),
             )
-
-    def add_org(self, reg, user):
-        from superset.savvy.organization import Organization
-        new_org = Organization()
-        new_org.organization_name = reg.organization
-        new_org.users.append(user)
-        try:
-            self.get_session.add(new_org)
-            self.get_session.commit()
-            return new_org
-        except Exception as e:
-            log.error(const.LOGMSG_ERR_SEC_ADD_REGISTER_USER.format(str(e)))
-            self.appbuilder.get_session.rollback()
-            return None
-
-    def add_register_user_org_admin(self, organization, first_name, last_name, email, password='', hashed_password=''):
-
-        register_user = self.registeruser_model()
-        register_user.first_name = first_name
-        register_user.last_name = last_name
-        register_user.email = email
-        register_user.organization = organization
-        if hashed_password:
-            register_user.password = hashed_password
-        else:
-            register_user.password = generate_password_hash(password)
-        register_user.registration_hash = str(uuid.uuid1())
-        try:
-            self.get_session.add(register_user)
-            self.get_session.commit()
-            return register_user
-        except Exception as e:
-            log.error(const.LOGMSG_ERR_SEC_ADD_REGISTER_USER.format(str(e)))
-            self.appbuilder.get_session.rollback()
-            return None
