+++ superset/utils/core.py
@@ -1081,8 +1081,3 @@ def MediumText():
 
 def shortid():
     return '{}'.format(uuid.uuid4())[-12:]
-
-
-def post_request(url, params):
-    import requests
-    return requests.post(url, data=json.dumps(params))
